p1-task4: completed

+++ include/os/task.h
@@ -16,7 +16,7 @@
 typedef struct {
     char name[16];
     int phyaddr_start, phyaddr_end;
-    int entrance;
+    uint64_t entrance;
 } task_info_t;
 
 extern task_info_t tasks[TASK_MAXNUM];
